Remove `PreparePlugin`, refactor `SyncPlugin` into `PhysicsTransformPlugin`, improve system sets
